Basic polyphony implemented. Notes can be turned on and off via uart. Some changes were built but not tested.

+++ main/synth_audio.h
@@ -4,7 +4,7 @@
 #define TIMER_FREQ 40000000 // 40 MHz
 
 
-uint32_t wave(uint32_t midi_note_number, uint32_t time);
+uint_fast32_t wave(uint_fast8_t midi_note_number, uint_fast32_t time);
 void task_audio_generate();
 void timer_init();
 void dac_init();
